+++ apps/web/pages/game/[room_id].tsx
@@ -1,27 +1,24 @@
 /** @format */
 
-import { IconArrowNarrowRight, IconDice } from "@tabler/icons";
 import { getPlayer } from "@troof/api";
 import { BadRequest, NotFoundResponse } from "@troof/responses";
 import {
 	Action,
 	EVENTS,
-	Log,
 	Player,
 	SECURITY_EVENTS,
 	Status,
 	TRUTH_OR_DARE_GAME,
 } from "@troof/socket";
 import { AxiosError, isAxiosError } from "axios";
-import classNames from "classnames";
 import { AnimatePresence, motion } from "framer-motion";
 import { NextPageContext } from "next";
 import Head from "next/head";
 import { useRouter } from "next/router";
 import { useContext, useEffect, useState } from "react";
-import { PropagateLoader } from "react-spinners";
 import Container from "../../components/Container";
 import FullScreenLoadingScreen from "../../components/FullScreenLoadingScreen";
+import MainItemSection from "../../components/game/MainItemSection";
 import RoomCodeSection from "../../components/game/RoomCodeSection";
 import ChatBox from "../../components/message/ChatBox";
 import EmojiReactionScreen from "../../components/message/EmojiReactionScreen";
@@ -153,7 +150,7 @@ export default function GamePage({
 					},
 				}}
 				transition={{ ease: "easeOut", type: "tween" }}
-				className="fixed left-0 top-0 z-[100] flex h-screen w-screen items-center justify-center bg-stone-300"
+				className="fixed left-0 top-0 z-[100] flex h-screen w-screen items-center justify-center bg-stone-400"
 			></motion.div>
 			<PublicKeyProvider>
 				<SocketProvider>
@@ -188,10 +185,8 @@ function GamePageContent({ r: roomID, player: p }: GamePageProps) {
 	const [gameStatus, setGameStatus] = useState<string>("in_lobby");
 
 	const [currentPlayer, setCurrentPlayer] = useState<Partial<Player>>({});
-	const [action, setAction] = useState<Action>(Action.Waiting_For_Selection);
 	const [text, setText] = useState<string>("");
-
-	const [isLoadingState, setLoadingState] = useState<boolean>(true);
+	const [action, setAction] = useState<Action>(Action.Waiting_For_Selection);
 
 	const [hasReceivedLatestLogItem, setHasReceivedLatestLogData] =
 		useState<boolean>(false);
@@ -229,33 +224,6 @@ function GamePageContent({ r: roomID, player: p }: GamePageProps) {
 				setGameStatus(data.status);
 				setHasReceivedGameStatus(true);
 			});
-
-			socket.on(TRUTH_OR_DARE_GAME.CONTINUE, (log: Log, player: Player) => {
-				console.log("Continue game received");
-				console.log(log, player);
-				setCurrentPlayer(player);
-				setText("");
-				setAction(log.action as Action);
-
-				setLoadingState(false);
-			});
-
-			socket.on(
-				TRUTH_OR_DARE_GAME.INCOMING_DATA,
-				(log: Log, player: Player) => {
-					console.log("New data received");
-					console.log(log, player);
-
-					setText(log.data);
-					setCurrentPlayer(player ?? {});
-					setAction(
-						(log.action as Action) ?? (Action.Waiting_For_Selection as Action)
-					);
-					setHasReceivedLatestLogData(true);
-
-					setLoadingState(false);
-				}
-			);
 
 			socket.on(SECURITY_EVENTS.PUBLIC_KEY, (publicKey: string) => {
 				setPublicKey(publicKey);
@@ -337,50 +305,6 @@ function GamePageContent({ r: roomID, player: p }: GamePageProps) {
 		}
 	}, [players, room_id, socket]);
 
-	const inLobbyGame = () => {
-		if (!socket) return;
-		socket.emit(EVENTS.GAME_UPDATE, {
-			room_id: room_id,
-			status: Status.In_Lobby,
-		});
-	};
-
-	const selectTruth = () => {
-		console.log("Selecting truth");
-		console.log(!!socket);
-
-		setLoadingState(true);
-		if (!socket) return;
-		console.log("Emitting to server to select truth");
-		socket.emit(TRUTH_OR_DARE_GAME.SELECT_TRUTH, {
-			room_id: room_id,
-		});
-	};
-
-	const selectDare = () => {
-		if (!socket) return;
-		setLoadingState(true);
-		socket.emit(TRUTH_OR_DARE_GAME.SELECT_DARE, {
-			room_id: room_id,
-		});
-	};
-
-	const handleContinue = () => {
-		if (!socket) return;
-		setLoadingState(true);
-		socket.emit(TRUTH_OR_DARE_GAME.CONTINUE, {
-			room_id: room_id,
-		});
-	};
-
-	const handleLeaveGame = () => {
-		if (!socket) return;
-		socket.emit(TRUTH_OR_DARE_GAME.LEAVE_GAME, {
-			room_id: room_id,
-			player_id: player.player_id,
-		});
-	};
-
 	return (
 		<Container>
 			<Head>
@@ -431,192 +355,20 @@ function GamePageContent({ r: roomID, player: p }: GamePageProps) {
 						</div>
 					</div>
 
+					{/* Main Content */}
 					<div className="col-span-2 rounded-2xl border-black/10 lg:h-full lg:border lg:px-10">
 						<div className="flex h-full w-full items-center justify-center">
-							{/* Main items */}
-							<div className=" my-2 flex-1">
-								{/* Current Player Name */}
-								<main className="my-10 flex w-full items-center justify-center">
-									<motion.div
-										className={`rnd bdr w-fit px-10 py-5 ${classNames({
-											"animate-zoom":
-												currentPlayer.player_id === player.player_id &&
-												action === Action.Waiting_For_Selection,
-										})}`}
-									>
-										<h1 className="break-all text-center font-Playfair text-5xl font-black">
-											{currentPlayer.display_name}
-										</h1>
-									</motion.div>
-								</main>
-
-								{/* The truth/dare box*/}
-								{action !== Action.Waiting_For_Selection && (
-									<motion.div
-										initial={{
-											opacity: 0,
-											y: -100,
-										}}
-										animate={{
-											opacity: 1,
-											y: 0,
-										}}
-										exit={{
-											opacity: 0,
-											y: 100,
-										}}
-									>
-										<div className="rnd bdr mx-auto my-3 w-fit space-y-10 px-10 py-5">
-											<p className="mb-5 text-center text-lg uppercase tracking-widest">
-												{action}
-											</p>
-
-											<h2 className="text-center text-xl font-semibold leading-normal md:text-3xl">
-												{text}
-											</h2>
-
-											{player.player_id === currentPlayer.player_id && (
-												<motion.button
-													whileHover={{
-														scale: 1.1,
-													}}
-													whileTap={{
-														scale: 0.9,
-													}}
-													className="mx-auto flex w-fit items-center justify-center gap-1 rounded-lg bg-black py-1 px-2 text-sm text-white disabled:text-opacity-50"
-													disabled={isLoadingState}
-													onClick={
-														action === Action.Dare
-															? () => {
-																	console.log("Re-rolling dare");
-																	selectDare();
-															  }
-															: () => {
-																	console.log("Re-rolling truth");
-																	selectTruth();
-															  }
-													}
-												>
-													<motion.div
-														key={Math.random()}
-														animate={{
-															rotate: 360,
-														}}
-													>
-														<IconDice size={19} />
-													</motion.div>
-													Re-roll
-												</motion.button>
-											)}
-										</div>
-									</motion.div>
-								)}
-
-								{/* If it is the current player and the action is to wait for a selection, Show the selection truth or dare buttons */}
-								{players.length >= 2 &&
-									currentPlayer.player_id === player.player_id &&
-									action === Action.Waiting_For_Selection && (
-										<motion.div>
-											<p className="text-center">Select One</p>
-
-											<div className="my-10 flex flex-wrap items-center justify-center gap-5">
-												<motion.button
-													whileHover={{
-														scale: 1.1,
-														y: -10,
-													}}
-													whileTap={{
-														scale: 0.9,
-													}}
-													className="btn-huge aspect-square w-[120px] disabled:text-opacity-50"
-													disabled={isLoadingState}
-													onClick={selectTruth}
-												>
-													Truth
-												</motion.button>
-
-												<motion.button
-													whileHover={{
-														scale: 1.1,
-														y: -10,
-													}}
-													whileTap={{
-														scale: 0.9,
-													}}
-													className="btn-huge aspect-square w-[120px] disabled:text-opacity-50"
-													disabled={isLoadingState}
-													onClick={selectDare}
-												>
-													Dare
-												</motion.button>
-											</div>
-										</motion.div>
-									)}
-
-								{players.length < 2 && (
-									<>
-										<p className="text-center italic">
-											Waiting for more players to join. (Min. 2)
-										</p>
-									</>
-								)}
-
-								{/* Show this below if the current player is not the player and that the action is waiting for selection */}
-								{currentPlayer.player_id !== player.player_id &&
-									action === Action.Waiting_For_Selection && (
-										<p className="text-center">
-											Waiting for {currentPlayer.display_name} to select
-										</p>
-									)}
-
-								{/* If it is the current player and they're not waiting for selection */}
-								{currentPlayer.player_id === player.player_id &&
-									action !== Action.Waiting_For_Selection && (
-										<div className="flex justify-center">
-											<motion.button
-												whileTap={{
-													scale: 0.9,
-												}}
-												initial={{ scale: 1 }}
-												animate={{
-													scale: 1.1,
-												}}
-												transition={{
-													repeat: Infinity,
-													repeatType: "mirror",
-													duration: 1,
-												}}
-												className="t my-1 mt-2 flex items-center gap-2 rounded-xl bg-green-300 px-5 py-3 text-sm font-semibold text-green-900 disabled:text-opacity-50"
-												onClick={handleContinue}
-												disabled={isLoadingState}
-											>
-												Continue
-												<motion.div
-													initial={{ x: 0 }}
-													animate={{
-														x: 10,
-													}}
-													transition={{
-														repeat: Infinity,
-														repeatType: "mirror",
-														duration: 1,
-													}}
-												>
-													<IconArrowNarrowRight size={16} />
-												</motion.div>
-											</motion.button>
-										</div>
-									)}
-
-								<div className="flex h-[50px] items-center justify-center">
-									<PropagateLoader
-										loading={isLoadingState}
-										color={"black"}
-										className="my-10 bg-red-500"
-										size={10}
-									/>
-								</div>
-							</div>
+							<MainItemSection
+								action={action}
+								room_id={roomID}
+								player={player}
+								players={players}
+								text={text}
+								currentPlayer={currentPlayer as Player}
+								setAction={setAction}
+								setCurrentPlayer={setCurrentPlayer}
+								setText={setText}
+							/>
 						</div>
 					</div>